Fix crash in OPENCL part The problem was type conversion:   cl_device_id -> long -> cl_device_id
On Windows long is 32 bit, but cl_device_id is 64 bit pointer
So, this conversion destroyed the cl_device_id

+++ src/plugins/smith_waterman/src/SWAlgorithmTask.cpp
@@ -238,7 +238,7 @@ void SWAlgorithmTask::prepare() {
         if(gpuMemBytes < needMemBytes) {
             stateInfo.setError(QString("Not enough memory on OpenCL-enabled device. "
                 "The space required is %1 bytes, but only %2 bytes are available. Device id: %3, device name: %4").
-                arg(QString::number(needMemBytes), QString::number(gpuMemBytes), QString::number(openClGpu->getId()), QString(openClGpu->getName())));
+                arg(QString::number(needMemBytes), QString::number(gpuMemBytes), QString::number((qlonglong)openClGpu->getId()), QString(openClGpu->getName())));
             return;
         } else {
             algoLog.details(QString("The Smith-Waterman search allocates ~%1 bytes (%2 Mb) on OpenCL device").
@@ -829,7 +829,7 @@ void PairwiseAlignmentSmithWatermanTask::prepare() {
         if(gpuMemBytes < needMemBytes) {
             stateInfo.setError(QString("Not enough memory on OpenCL-enabled device. "
                 "The space required is %1 bytes, but only %2 bytes are available. Device id: %3, device name: %4").
-                arg(QString::number(needMemBytes), QString::number(gpuMemBytes), QString::number(openClGpu->getId()), QString(openClGpu->getName())));
+                arg(QString::number(needMemBytes), QString::number(gpuMemBytes), QString::number((qlonglong)openClGpu->getId()), QString(openClGpu->getName())));
             return;
         } else {
             algoLog.details(QString("The Smith-Waterman search allocates ~%1 bytes (%2 Mb) on OpenCL device").
